BUG: default_index creates int64 even in i32-native platforms

+++ pandas/core/common.py
@@ -587,7 +587,7 @@ def _is_bool_indexer(key):
 
 def _default_index(n):
     from pandas.core.index import Int64Index
-    values = np.arange(n)
+    values = np.arange(n, dtype=np.int64)
     result = values.view(Int64Index)
     result.name = None
     return result
